ai-chat-backend/requirements.txt
flask
flask-cors
python-dotenv
werkzeug
PyJWT
requests
